fit.model

Fit negative binomial model to obtain p-values for interactions.

Usage

fit.model(
  linkSet,
  distance.bins = NULL,
  distribution = "negative-binomial",
  adjustment.terms = NULL,
  maxit = 100,
  epsilon = 1e-08,
  cores = 1,
  trace = FALSE,
  verbose = FALSE,
  interim.data.dir = NULL
)

Arguments

distribution

Name of distribution of the counts. Options are 'negative-binomial', 'poisson', 'truncated-poisson', and 'truncated-negative-binomial'

adjustment.terms

Character vector of extra terms to adjust for in the model fit.

maxit

Maximum number of IWLS iterations for fitting the model (passed to glm.control)

epsilon

Positive convergence tolerance for Poisson and negative binomial models. Passed to glm.control

cores

Integer value specifying how many cores to use to fit model for cis-interactions.

trace

Logical indicating if output should be produced for each of model fitting procedure. Passed to glm.control or gamlss.control

verbose

Logical indicating whether to print progress reports.

interim.data.dir

Path to directory to store intermediate QC data and plots.

interaction.data

data.table object containing interaction counts. Must contain columns distance, count, and bait_trans_count.

Value

Interactions data with expected number of interactions and p-values added.

Details

Fit a negative binomial model for obtaining p-value for interactions. The data is first sorted by distance, and models
are fit separately in each quantile of the distance-sorted data.

Examples

# \donttest{
  data(bre80);
#> Warning: data set ‘bre80’ not found
fit.model(bre80);
#> [1] "fitting model...."
#> Error: object 'bre80' not found
# }