linkSet-accessors

Methods to get and set fields in an linkSet object.

This method returns the anchor IDs of a linkSet object.

This method returns the anchors of a linkSet object.

This method returns the bait anchors of a linkSet object.

This method returns the other end (oe) anchors of a linkSet object.

This method is an alias for 'first' and returns the bait anchors of a linkSet object.

This method is an alias for 'second' and returns the other end (oe) anchors of a linkSet object.

This method returns the regions of a linkSet object.

This method returns the regions corresponding to the bait anchors of a linkSet object.

This method replaces the bait anchors of a linkSet object with new values.

This method replaces the regions of a linkSet object with new values.

This method replaces the anchor1 of a linkSet object with new values.

This method replaces the anchor2 of a linkSet object with new values.

This method replaces the regions of a linkSet object with new values.

This method replaces the regions corresponding to the bait anchors of a linkSet object with new values.

This method replaces the other end (oe) anchors of a linkSet object with new values.

This method returns the metadata column of a linkSet object.

This method replaces the metadata column of a linkSet object with new values.

Usage

bait(x) <- value

regions(x) <- value

anchor1(x) <- value

anchor2(x) <- value

unchecked_regions(x) <- value

# S4 method for class 'linkSet'
anchor1(x)

# S4 method for class 'linkSet'
anchor2(x)

# S4 method for class 'linkSet'
regions(x)

# S4 method for class 'linkSet'
seqinfo(x)

# S4 method for class 'linkSet'
anchorIds(x, type = "both")

# S4 method for class 'linkSet'
anchors(x, type = "both", id = FALSE)

# S4 method for class 'linkSet'
first(x)

# S4 method for class 'linkSet'
second(x)

# S4 method for class 'linkSet'
bait(x)

# S4 method for class 'linkSet'
oe(x)

# S4 method for class 'linkSet'
regions(x)

# S4 method for class 'linkSet'
regionsBait(x)

# S4 method for class 'linkSet'
bait(x) <- value

# S4 method for class 'linkSet'
unchecked_regions(x) <- value

# S4 method for class 'linkSet'
unchecked_anchor1(x) <- value

# S4 method for class 'linkSet'
unchecked_anchor2(x) <- value

# S4 method for class 'linkSet'
regions(x) <- value

# S4 method for class 'linkSet'
regionsBait(x) <- value

# S4 method for class 'linkSet'
oe(x) <- value

# S4 method for class 'linkSet'
x$name

# S4 method for class 'linkSet'
x$name <- value

Arguments

x

A linkSet object.

value

The new value to assign to the specified metadata column.

type

The type of anchor to return. Can be "both", "bait", or "oe".

id

If TRUE, returns the anchor IDs instead of the anchors.

name

A character string specifying the name of the metadata column to replace.

Value

For the getters, values in various slots of x are returned, while for the setters, the slots of x are modified accordingly – see Details.

A vector of the regions

A list of anchor IDs.

A list of anchors or anchor IDs.

A GRanges object containing the bait anchors.

A GRanges object containing the oe anchors.

A GRanges object containing the bait anchors.

A GRanges object containing the oe anchors.

A GRanges object containing the regions.

A GRanges object containing the regions corresponding to the bait anchors.

The modified linkSet object with the new bait anchors.

The modified linkSet object with the new regions.

The modified linkSet object with the new anchor1 values.

The modified linkSet object with the new anchor2 values.

The modified linkSet object with the new regions.

The modified linkSet object with the new regions corresponding to the bait anchors.

The modified linkSet object with the new other end anchors.

The value of the specified metadata column.

The modified linkSet object with the new metadata column value.

Author

Gilbert Han

Examples

data(linkExample)
anchor1(linkExample)
#> [1] 1 3 5 7 9
data(linkExample)
anchor2(linkExample)
#> [1]  2  4  6  8 10
data(linkExample)
regions(linkExample)
#> Error: unable to find an inherited method for function ‘regions’ for signature ‘x = "linkSet"’
data(linkExample)
anchorIds(linkExample, type="both")
#> Error: unable to find an inherited method for function ‘anchorIds’ for signature ‘x = "linkSet"’
data(linkExample)
anchors(linkExample, type="both", id=FALSE)
#> Error: unable to find an inherited method for function ‘anchors’ for signature ‘x = "linkSet"’