Plot Baits

Plot baits in a linkSet object

Usage

plotBaits(
  linkset,
  scoreCol = "score",
  countCol = "count",
  n = 4,
  baits = NULL,
  plotBaitNames = TRUE,
  plevel1 = 5,
  plevel2 = 3,
  outfile = NULL,
  width = 20,
  height = 20,
  extend.base = 1e+06,
  bgCol = "black",
  lev2Col = "blue",
  lev1Col = "red",
  ...
)

Arguments

linkset

A linkSet object

scoreCol

Column name containing scores for coloring points

countCol

Column name containing counts for y-axis values

n

Number of random baits to plot if baits parameter is NULL

baits

Vector of specific baits to plot. If NULL, n random baits are selected

plotBaitNames

Logical indicating whether to show bait names in plot titles

plevel1

Upper threshold for score coloring (red)

plevel2

Lower threshold for score coloring (blue)

outfile

Output file path. If NULL, plot is displayed rather than saved

width

Width of output plot in inches

height

Height of output plot in inches

extend.base

Base pairs to extend view range on either side of bait

bgCol

Color for points below plevel2 threshold

lev2Col

Color for points between plevel2 and plevel1 thresholds

lev1Col

Color for points above plevel1 threshold

...

Additional plotting parameters

Value

A ggplot object