run_chicane

This function adapts the chicane function from the ChICANE package to work with the linkSet object format. Run full method for detecting significant interactions in capture Hi-C experiments, starting either from a linkSet object or preprocessed data from prepare.data

Usage

# S4 method for class 'linkSet'
run_chicane(
  linkSet,
  replicate.merging.method = "sum",
  distribution = "negative-binomial",
  include.zeros = "none",
  bait.filters = c(0, 1),
  target.filters = c(0, 1),
  distance.bins = NULL,
  multiple.testing.correction = c("bait-level", "global"),
  adjustment.terms = NULL,
  remove.adjacent = FALSE,
  temp.directory = NULL,
  keep.files = FALSE,
  maxit = 100,
  epsilon = 1e-08,
  cores = 1,
  trace = FALSE,
  verbose = FALSE,
  interim.data.dir = NULL
)

Arguments

linkSet

A linkSet object containing interaction data. Can be used instead of interactions specification if the linkSet object has already been prepared.

replicate.merging.method

Method that should be used for merging replicates, if applicable

distribution

Name of distribution of the counts. Options are 'negative-binomial', 'poisson', 'truncated-poisson', and 'truncated-negative-binomial'

bait.filters

Vector of length two, where the first element corresponds to the lower-end filter and the second to the upper-end filter. When global multiple testing correction is performed, altering the bait filtering settings may affect the number of significant results.

target.filters

Vector of length two, giving lower and higher filter, respectively. Changing this filtering setting may affect multiple testing correction by altering the number of tests performed.

distance.bins

Number of bins to split distance into. Models are fit separately in each bin.

multiple.testing.correction

String specifying how multiple testing correction should be performed, by bait or globally.

adjustment.terms

Character vector of extra terms to adjust for in the model fit.

maxit

Maximum number of IWLS iterations for fitting the model (passed to glm.control)

epsilon

Positive convergence tolerance for Poisson and negative binomial models. Passed to glm.control

cores

Integer value specifying how many cores to use to fit model for cis-interactions.

trace

Logical indicating if output should be produced for each of model fitting procedure. Passed to glm.control or gamlss.control

verbose

Logical indicating whether to print progress reports.

interim.data.dir

Path to directory to store intermediate QC data and plots. NULL indicate skip intermediate results.

Value

A linkSet object with additional columns:

expected

The expected number of reads linking the two fragments under the fitted model

p.value

P-value for test of the observed number of reads significantly exceeding the expected count

q.value

FDR-corrected p-value