run_chicane

This function adapts the chicane function from the ChICANE package to work with the linkSet object format. Run full method for detecting significant interactions in capture Hi-C experiments, starting either from a linkSet object or preprocessed data from prepare.data

Verify that linkSet object is in expected format. Throws an error if object does not fit requirements.

Check if an error matches the error raised by \code{glm.nb} due to an inflated theta estimate.

This happens when the variance of the negative binomial does not exceed the mean (i.e. there is no overdispersion). In such cases, the Poisson distribution may be a suitable alternative.

Check if chicane model can be fit on a given dataset.

glm.nb does not work when all responses are constant, or there are only two unique values and a covariate is a perfect predictor.

Perform multiple testing correction on p-values from interaction test. By default, multiple testing correction is applied per bait. To change this to a global multiple testing correction, set bait.level = FALSE.

Split a data frame into a prespecified number of bins, using split and cut. Unlike the default R functions, this does not fail when asked to split the data into a single bin.

Verify that interaction.data object is in expected format. Throws an error if object does not fit requirements.

Fit GLM according to a specified distribution. This needs to be done separately from glm in order to include negative binomial and truncated distributions as options.

Check that the model fit contains the same number of rows as the data used to fit it,

and throw an error if not

Check if a warning object is an iteration limit reached warning from glm.nb

Check if a warning matches the square root warning raised by \code{glm.nb} due to an inflated theta estimate.

This happens when the variance of the negative binomial does not exceed the mean (i.e. there is no overdispersion). In such cases, the Poisson distribution may be a suitable alternative.

Usage

# S4 method for class 'linkSet'
run_chicane(
  linkSet,
  replicate.merging.method = "sum",
  distribution = "negative-binomial",
  include.zeros = "none",
  bait.filters = c(0, 1),
  target.filters = c(0, 1),
  distance.bins = NULL,
  multiple.testing.correction = c("bait-level", "global"),
  adjustment.terms = NULL,
  remove.adjacent = FALSE,
  temp.directory = NULL,
  keep.files = FALSE,
  maxit = 100,
  epsilon = 1e-08,
  cores = 1,
  trace = FALSE,
  verbose = FALSE
)

.verify.linkSet(linkSet)

.is.glm.nb.theta.error(e)

.check.model.numerical.fit(interaction.data)

multiple.testing.correct(interaction.data, bait.level = TRUE)

.smart.split(dat, bins)

.verify.interaction.data(interaction.data)

fit.glm(
  formula,
  data,
  distribution = c("negative-binomial", "poisson", "truncated-poisson",
    "truncated-negative-binomial"),
  start = NULL,
  init.theta = NULL,
  maxit = 100,
  epsilon = 1e-08,
  trace = FALSE
)

.model.rows.sanity.check(model.data, model)

.is.glm.nb.maxiter.warning(w)

.is.glm.nb.theta.warning(w)

Arguments

linkSet

Object to be verified.

replicate.merging.method

Method that should be used for merging replicates, if applicable

distribution

Name of distribution of the counts. Options are 'negative-binomial', 'poisson', 'truncated-poisson', and 'truncated-negative-binomial'

bait.filters

Vector of length two, where the first element corresponds to the lower-end filter and the second to the upper-end filter. When global multiple testing correction is performed, altering the bait filtering settings may affect the number of significant results.

target.filters

Vector of length two, giving lower and higher filter, respectively. Changing this filtering setting may affect multiple testing correction by altering the number of tests performed.

distance.bins

Number of bins to split distance into. Models are fit separately in each bin.

multiple.testing.correction

String specifying how multiple testing correction should be performed, by bait or globally.

maxit

Maximum number of IWLS iterations for fitting the model (passed to glm.control)

epsilon

Positive convergence tolerance for Poisson and negative binomial models. Passed to glm.control

trace

Logical indicating if output should be produced for each of model fitting procedure. Passed to glm.control or gamlss.control

verbose

Logical indicating whether to print progress reports.

e

Error object

interaction.data

Object to be verified.

bait.level

Logical indicating whether multiple testing correction should be performed per bait.

dat

Data frame or data table to be split

bins

Number of bins to split data into

formula

Formula specifying model of interest

data

Data frame containing variables specified in formula

start

Starting values for model coefficients

init.theta

Initial value of theta if fitting the negative binomial distribution

model.data

Data used to fit model

model

Resulting negative binomial model object

w

Warning object

Value

A linkSet object with additional columns:

expected

The expected number of reads linking the two fragments under the fitted model

p.value

P-value for test of the observed number of reads significantly exceeding the expected count

q.value

FDR-corrected p-value

None

Boolean indicating if error matches

boolean indicating if model can be fit

Original data table with new column

q.value

FDR-corrected p-value

List with bins elements. Each element corresponds to one portion of the data

None

List with elements

model

model object

expected.values

vector of expected values for each element in original data

p.values

vector of p-values for test of significantly higher response than expected

None

Logical indicating if warning matches iteration limit reached warning

Boolean indicating if warning matches

Examples

# Example usage of run_chicane function
gr1 <- GRanges(seqnames = c("chr1", "chr3", "chr3"),
               ranges = IRanges(start = c(1000, 2000, 3000), width = 100),
               strand = "+", symbol = c("BRCA1", "TP53", "NONEXISTENT"))
gr2 <- GRanges(seqnames = c("chr1", "chr2", "chr3"),
               ranges = IRanges(start = c(5000, 6000, 7000), width = 100),
               strand = "+")
ls <- linkSet(gr1, gr2, specificCol = "symbol")
annotated_ls <- suppressWarnings(annotatePromoter(ls, genome = "hg38", upstream = 500,overwrite = TRUE))
annotated_ls <- countInteractibility(annotated_ls)
annotated_ls <- linkSet::pairdist(annotated_ls)
# Run run_chicane function
result_ls <- run_chicane(annotated_ls, replicate.merging.method = 'sum', 
                         bait.filters = c(0, 1), target.filters = c(0, 1), 
                         distance.bins = NULL, multiple.testing.correction = 'bait-level', 
                         verbose = TRUE)
#> [1] "Not found column 'bait.id', adding bait name as default."
#> [1] "Not found column 'bait.to.bait' Set 'FALSE' as default."
#> FITTING MODEL
#> [1] "fitting model...."
#> 
#> Attaching package: ‘data.table’
#> The following object is masked from ‘package:SummarizedExperiment’:
#> 
#>     shift
#> The following object is masked from ‘package:GenomicRanges’:
#> 
#>     shift
#> The following object is masked from ‘package:IRanges’:
#> 
#>     shift
#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, second
#> *
#> Warning: no observations informative at iteration 1
#> Warning: no observations informative at iteration 1
#> Warning: glm.fit: algorithm did not converge
#> 
#> 	trans interactions
result_ls  
#> linkSet object with 3 interactions and 11 metadata columns:
#>              bait     seqnames_oe ranges_oe | anchor1.symbol  inter_type
#>       <character>           <Rle> <IRanges> |    <character> <character>
#>   [1]       BRCA1 ---        chr1 5000-5099 |          BRCA1       intra
#>   [2]        TP53 ---        chr2 6000-6099 |           TP53       intra
#>   [3] NONEXISTENT ---        chr3 7000-7099 |    NONEXISTENT       intra
#>           count bait.trans.count target.trans.count  distance     bait.id
#>       <integer>        <numeric>          <numeric> <integer> <character>
#>   [1]         1                1                  1      <NA>       BRCA1
#>   [2]         1                1                  1      <NA>        TP53
#>   [3]         1                1                  1      <NA> NONEXISTENT
#>       bait.to.bait  expected   p.value   q.value
#>          <logical> <numeric> <numeric> <numeric>
#>   [1]        FALSE         1  0.632121  0.632121
#>   [2]        FALSE         1  0.632121  0.632121
#>   [3]        FALSE         1  0.632121  0.632121
#>   -------
#>   regions: 6 ranges and 0 metadata columns
#>   seqinfo: 712 sequences (1 circular) from 2 genomes (hg38, NA)